pyproject.toml
[project]
name = "api_key_verifier"
version = "0.1.0"
description = "A tool to verify Google Gemini API keys."
authors = [
    { name = "Maher", email = "your.email@example.com" }
]
dependencies = [
    "google-generativeai~=0.3.0",
    "python-dotenv~=1.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }

[project.scripts]
verify-key = "api_key_verifier.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
